moved env check back again

+++ src/api/common/helpers/seed-database.js
@@ -2,6 +2,10 @@ import models from '~/src/api/common/models/index.js'
 import data from '~/src/api/common/helpers/sample-data/index.js'
 
 export async function seedDatabase(db, logger) {
+  if (process.env.NODE_ENV === 'production' || process.env.JEST_WORKER_ID) {
+    return
+  }
+
   for (const [name, model] of Object.entries(models)) {
     try {
       await model.db.dropCollection(name)
